GRIDEDIT-1869: pull remote release branch
When making a patch release with cherry-picked changes, the release
branch already exists, and the starting point is the same as the version
number. In that scenario the remote branch needs to be pulled.

Note that this change only does the pull, but still first creates the
changes on the local branch. An improvement would be to first detect the
existence of the release branch and pull it, and not make any local
changes.